requirements.txt
discord.py==1.7.3
googlesearch-python==1.0.1
python-decouple==3.5